+++ src/app/page.tsx
@@ -2,9 +2,48 @@ import { z } from 'zod';
 
 import { DiGithubBadge } from 'react-icons/di';
 
+import { Metadata } from 'next';
+
 import DtConfigCard from '@/components/features/dtConfigCard';
 import DtConvertCard from '@/components/features/dtConvertCard';
 import { ThemeToggle } from '@/components/theme-toggle';
+
+export const runtime = 'edge';
+
+export async function generateMetadata({
+  searchParams,
+}: {
+  searchParams: Promise<{ dt: string | undefined }>;
+}): Promise<Metadata> {
+  const { dt } = await searchParams;
+
+  const baseUrl = new URL(
+    process.env.NEXT_PUBLIC_BASE_URL || 'https://zonr.dev',
+  );
+
+  const metadata: Metadata = {
+    title: 'zonr - Share your Time',
+    metadataBase: baseUrl,
+    description:
+      'zonr is a platform for sharing your time with others in different timezones',
+    openGraph: {
+      title: 'zonr - Share your Time',
+      description:
+        'zonr is a platform for sharing your time with others in different timezones',
+      images: dt
+        ? [`/opengraph-image?dt=${encodeURIComponent(dt)}`]
+        : [`/opengraph-image`],
+    },
+    twitter: {
+      card: 'summary_large_image',
+      images: dt
+        ? [`/opengraph-image?dt=${encodeURIComponent(dt)}`]
+        : [`/opengraph-image`],
+    },
+  };
+
+  return metadata;
+}
 
 function isValidDateTime(dt: string) {
   try {
